requirements.txt
pdf2image
pytesseract
PyMuPDF
Pillow
opencv-python
pytest
python-dotenv
openai